+++ app/Http/Controllers/RequiredPartRequisitionController.php
@@ -157,9 +157,16 @@ class RequiredPartRequisitionController extends Controller
      */
     public function destroy($id)
     {
-        $requisition = RequiredPartRequisition::find($id)->delete();
-        RequiredPartItems::where('model_id',$id)->delete();
-            return message('Requisition deleted successfully');
+        $requisition = RequiredPartRequisition::find($id);
+        if($requisition){
+            $requisition->delete();
+            RequiredPartItems::where('model_id',$id)->delete();
+                return message('Requisition deleted successfully',201);
+        }else{
+            return message('Requisition deleted successfully',422);
+
+        }
+        
     }
 
     public function RequiredRequisitionStatus(Request $request, $id)
